third_party/WebKit/LayoutTests/fast/html/imports/import-mode-quirks.html

<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link rel="import" id=withDTD href="resources/hello.html">
<link rel="import" id=noDTD href="resources/hello-nodtd.html">
<link rel="import" id=legacyDTD href="resources/hello-legacy.html">
</head>
<body>
<script>
test(function(){
  assert_equals(document.compatMode, "BackCompat");
  assert_equals(withDTD.import.compatMode, "CSS1Compat");
  assert_equals(noDTD.import.compatMode, "CSS1Compat");
  assert_equals(legacyDTD.import.compatMode, "CSS1Compat");
}, "Each HTML Import is standard mode regardless of the mode of the master document.");
</script>
</body>
</html>
